roi_tools.py from utilities.py to avoid repeating code, remove dicom tools directory, rename dicom tool mlc_analyzer.py
